一键发布

+++ monster-admin-front/src/views/VideoView.vue
@@ -1,94 +1,105 @@
 <template>
-  <!-- <a-button class="upload-btn" type="primary" ghost>+ 上传视频</a-button> -->
-  <AddVedioModal :refresh="fetchVideoList" />
-  <a-table
-    :columns="columns"
-    :data-source="videoList"
-    :loading="loading"
-    :scroll="{ x: 800, y: 500 }"
-  >
-    <template #headerCell="{ column }">
-      <template v-if="column.key === 'id'">
-        <span> Id </span>
+  <div class="table-container">
+    <div class="addVideoBtn">
+      <AddVedioModal :refresh="fetchVideoList" />
+    </div>
+    <a-table
+      :columns="columns"
+      :data-source="videoList"
+      :loading="loading"
+      :scroll="{ x: 1000, y: 500 }"
+    >
+      <template #headerCell="{ column }">
+        <template v-if="column.key === 'id'">
+          <span> Id </span>
+        </template>
       </template>
-    </template>
-    <template #bodyCell="{ column, record }">
-      <template v-if="column.key === 'id'">
-        <a-button type="primary" @click="handleVideoDetail(record.id)" ghost>
-          {{ record.id }}
-        </a-button>
-      </template>
-      <template v-else-if="column.key === 'title'">
-        <span>
-          {{ record.title }}
-        </span>
-      </template>
-      <template v-else-if="column.key === 'url'">
-        <span>
-          <a :href="record.url" target="_blank">{{ record.url }}</a>
-        </span>
-      </template>
-      <template v-else-if="column.key === 'action'">
-        <a-button @click="handleDeleteVideo(record.id)" type="primary" danger>删除</a-button>
-      </template>
-    </template>
-  </a-table>
+      <template #bodyCell="{ column, record }">
+        <template v-if="column.key === 'id'">
+          <a-button type="primary" @click="handleVideoDetail(record.id)" ghost>
+            {{ record.id }}
+          </a-button>
+        </template>
+        <template v-else-if="column.key === 'title'">
+          <span>
+            {{ record.title }}
+          </span>
+        </template>
+        <template v-else-if="column.key === 'url'">
+          <span>
+            <a :href="record.url" target="_blank">{{ record.url }}</a>
+          </span>
+        </template>
+        <template v-else-if="column.key === 'action'">
+          <div class="actionBtn">
+            <!-- 一键上传 -->
+            <!-- <a-button @click="handlePublishVideo(record.id)" type="primary" ghost>
+              一键上传
+            </a-button> -->
+            <PublishVideoModal :record="record" />
 
-  <!-- 添加视频详情modal框 -->
-  <a-modal v-model:open="detailModalVisible" title="视频详情" :footer="null">
-    <template v-if="currentVideo">
-      <video :src="currentVideo.url" controls style="width: 100%; margin-bottom: 10px"></video>
-      <div>
-        <p>视频ID: {{ currentVideo.id }}</p>
-        <p>视频标题: {{ currentVideo.title }}</p>
-        <p>视频URL:</p>
-        <a :href="currentVideo.url" target="_blank" style="color: #1890ff">{{
-          currentVideo.url
-        }}</a>
-        <p></p>
-        <p>视频文件名: {{ currentVideo.filename }}</p>
-        <p>视频上传时间: {{ currentVideo.createdAt }}</p>
-        <p>视频修改时间: {{ currentVideo.updatedAt }}</p>
-      </div>
-    </template>
-  </a-modal>
+            <a-button @click="handleDeleteVideo(record.id)" type="primary" danger>删除</a-button>
+          </div>
+        </template>
+      </template>
+    </a-table>
+
+    <!-- 添加视频详情modal框 -->
+    <a-modal v-model:open="detailModalVisible" title="视频详情" :footer="null">
+      <template v-if="currentVideo">
+        <video :src="currentVideo.url" controls style="width: 100%; margin-bottom: 10px"></video>
+        <div>
+          <p>视频ID: {{ currentVideo.id }}</p>
+          <p>视频标题: {{ currentVideo.title }}</p>
+          <p>视频URL:</p>
+          <a :href="currentVideo.url" target="_blank" style="color: #1890ff">{{
+            currentVideo.url
+          }}</a>
+          <p></p>
+          <p>视频文件名: {{ currentVideo.filename }}</p>
+          <p>视频上传时间: {{ currentVideo.createdAt }}</p>
+          <p>视频修改时间: {{ currentVideo.updatedAt }}</p>
+        </div>
+      </template>
+    </a-modal>
+  </div>
 </template>
 
 <script lang="ts" setup>
-import { ref, onMounted } from 'vue'
+import { ref, onMounted, createVNode } from 'vue'
 import { getVideoList } from '@/api/user'
-import { message } from 'ant-design-vue'
+import { message, Modal } from 'ant-design-vue'
 import AddVedioModal from '@/components/AddVedioModal.vue'
 import { deleteVideo } from '@/api/user'
 import myAxios from '@/request'
+import { ExclamationCircleOutlined } from '@ant-design/icons-vue'
+import PublishVideoModal from '@/components/PublishVideoModal.vue'
 
 const columns = [
   {
-    name: 'Id',
+    title: 'Id',
     dataIndex: 'id',
     key: 'id',
     fixed: 'left',
-    width: 20,
+    width: 200,
   },
   {
     title: 'Title',
     dataIndex: 'title',
     key: 'title',
-    fixed: 'left',
-    width: 80,
+    width: 200,
   },
   {
     title: 'URL',
     dataIndex: 'url',
     key: 'url',
-    fixed: 'left',
-    width: 100,
+    width: 300,
   },
   {
     title: 'Action',
     key: 'action',
     fixed: 'right',
-    width: 30,
+    width: 100,
   },
 ]
 
@@ -108,11 +119,7 @@ async function fetchVideoList() {
     const res = await getVideoList()
     if (res.code === 0) {
       videoList.value = res.data || []
-    } else {
-      message.error('获取视频列表失败')
     }
-  } catch (error) {
-    message.error('网络请求失败')
   } finally {
     loading.value = false
   }
@@ -121,37 +128,52 @@ async function fetchVideoList() {
 // 删除视频
 async function handleDeleteVideo(id: string) {
   try {
-    const res = await deleteVideo(id)
-    if (res.code === 0) {
-      message.success('删除成功')
-      await fetchVideoList()
-    } else {
-      message.error(res.msg || '删除失败')
-    }
-  } catch (error) {
-    message.error('删除失败')
-  }
+    Modal.confirm({
+      title: '确认要删除该视频吗？',
+      icon: createVNode(ExclamationCircleOutlined),
+      content: '点击OK按钮后, 该视频将被删除!!',
+      onOk: async () => {
+        const res = await deleteVideo(id)
+        if (res.code === 0) {
+          await fetchVideoList()
+          message.success('删除成功')
+        }
+      },
+      onCancel() {},
+    })
+  } catch (error) {}
 }
 
 // 查看视频详情
 async function handleVideoDetail(id: string) {
-  const res = await myAxios.get(`/api/video/${id}`)
   try {
+    const res = await myAxios.get(`/api/video/${id}`)
     if (res.code === 0) {
       console.log(res.data)
       detailModalVisible.value = true
       currentVideo.value = res.data
-    } else {
-      message.error('获取视频详情失败', res.msg)
     }
-  } catch (error) {
-    message.error('获取视频详情失败', error)
-  }
+  } catch (error) {}
 }
+
+// 一键上传
+// async function handlePublishVideo(id: string) {
+//   try {
+//     const res = await myAxios.post(`/api/video/${id}`)
+//     if (res.code === 0) {
+//       message.success('上传成功')
+//     }
+//   } catch (error) {}
+// }
 </script>
 
 <style scoped>
-.upload-btn {
+.table-container {
+  width: 100%;
+  overflow: auto;
+}
+
+.addVideoBtn {
   margin-bottom: 10px;
 }
 
@@ -161,5 +183,15 @@ a {
 
 a:hover {
   color: #1890ff;
+}
+
+.actionBtn {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  width: 60%;
+  align-items: center;
+  justify-content: center;
+  margin-left: 20%;
 }
 </style>
